WIP: src/components/routes: add component RouteCalendarPanel to log data about active routes (#507)
* boilerplate for bottom panel

* panel boilerplate

* add route input components + fix title

* add route input components + fix title

* use panel + refactor route logging behaviour into composable

* calendar panel docs

* calendar panel add types

* refactor types

* route item edit - use route initial data

* item edit - fix new routeInitial usage

* calendar panel add tests

* calendar panel save

* refactor route input data

* load panel input data from routes

* clean up route logging logic

* use enum for transport direction

* fix routes calendar tests

* add routes calendar e2e test

* routes calendar test panel with single and multiple selected routes

* extend tests + docs

* cleanup panel tests

* panel heading styling and test

* distance validation - allow decimal points

* routes calendar - extend tests

* cleanup tests

* calendar panel docs

* variable rename

* replace transport type with enum

* variable rename

---------

+++ src/composables/useValidation.ts
@@ -62,7 +62,7 @@ export const useValidation = () => {
     if (typeof val === 'number') {
       return val > 0;
     } else {
-      const numberVal: number = parseInt(val);
+      const numberVal: number = parseFloat(val);
       if (isNaN(numberVal)) {
         return val?.length > 0;
       } else {
